refactor content item example block
